hackathon-templates/pyproject.toml
[project]
name = "hackathon-agent"
version = "0.1.0"
description = "AI Agent for Cloud Run Hackathon"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.13"
dependencies = [
    "adk-python>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "google-cloud-logging>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 